Implemented a trivial frequent itemset finder.
Given a dataset of sequences, a min support, a min size and a max size it can find itemset having a high enough support with a length in [min size, max size].
The algorithm is properly working but can easily be enhanced.

Suggestion: use a known frequent itemset mining algorithm.